requirements.txt
openai-whisper
pyaudio
wave
numpy
requests
munch
transfromers
torch
scipy